agents/requirements.txt
# Agents service dependencies
# Vertex AI via ADC — no API key needed
google-cloud-aiplatform>=1.49.0     # vertexai SDK + Model Garden access
google-adk>=1.0.0                   # Google Agent Development Kit
mcp>=1.3.0                          # Model Context Protocol client

# API + async
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
pydantic>=2.11.0
httpx>=0.28.1

# Database
sqlalchemy[asyncio]>=2.0.29
asyncpg>=0.29.0
pgvector>=0.2.5

# Observability + reliability
structlog>=24.1.0
tenacity>=8.3.0

# Secrets
google-cloud-secret-manager>=2.20.0
